Facade implemented, utils added
