+++ src/main/java/com/demoproject/cricketapp/controller/MatchController.java
@@ -1,6 +1,6 @@
 package com.demoproject.cricketapp.controller;
 
-import com.demoproject.cricketapp.service.impl.MatchServiceImpl;
+import com.demoproject.cricketapp.service.MatchService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("match")
 @RequiredArgsConstructor
 public class MatchController {
-    private final MatchServiceImpl matchServiceImpl;
+    private final MatchService matchService;
 
 //    @PostMapping("")
 //    public ResponseEntity<Match> createMatch(@RequestBody MatchRequest matchRequest) {
