+++ src/main/java/com/demoproject/cricketapp/controller/PlayerController.java
@@ -3,8 +3,7 @@ package com.demoproject.cricketapp.controller;
 import com.demoproject.cricketapp.beans.Player;
 import com.demoproject.cricketapp.beans.request.PlayerRequest;
 import com.demoproject.cricketapp.beans.response.PlayerInfoResponse;
-import com.demoproject.cricketapp.facade.CricketServicesFacade;
-import com.demoproject.cricketapp.service.PlayerService;
+import com.demoproject.cricketapp.facade.PlayerTeamFacade;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,18 +15,18 @@ import java.util.List;
 @RequestMapping("player")
 @RequiredArgsConstructor
 public class PlayerController {
-    private final CricketServicesFacade cricketServicesFacade;
+    private final PlayerTeamFacade playerTeamFacade;
     @PostMapping("")
     public ResponseEntity<Player> addPlayer(@RequestBody PlayerRequest playerRequest) {
-        return new ResponseEntity<Player>(cricketServicesFacade.addPlayer(playerRequest), HttpStatus.OK);
+        return new ResponseEntity<Player>(playerTeamFacade.addPlayer(playerRequest), HttpStatus.OK);
     }
     @GetMapping("")
     public ResponseEntity<List<PlayerInfoResponse>> getAllPlayersInfo() {
-        return new ResponseEntity<List<PlayerInfoResponse>>(cricketServicesFacade.getAllPlayersInfo(), HttpStatus.OK);
+        return new ResponseEntity<List<PlayerInfoResponse>>(playerTeamFacade.getAllPlayersInfo(), HttpStatus.OK);
     }
     @GetMapping("{id}")
     public ResponseEntity<Player> getPlayerById(@PathVariable String id) {
-        return new ResponseEntity<Player>(cricketServicesFacade.getPlayerById(id), HttpStatus.OK);
+        return new ResponseEntity<Player>(playerTeamFacade.getPlayerById(id), HttpStatus.OK);
     }
 
 }
