+++ src/main/java/com/demoproject/cricketapp/facade/PlayerTeamFacade.java
@@ -1,18 +1,18 @@
 package com.demoproject.cricketapp.facade;
 
 
-import com.demoproject.cricketapp.beans.Match;
 import com.demoproject.cricketapp.beans.Player;
-import com.demoproject.cricketapp.beans.Scoreboard;
 import com.demoproject.cricketapp.beans.Team;
 import com.demoproject.cricketapp.beans.request.PlayerRequest;
 import com.demoproject.cricketapp.beans.request.TeamRequest;
 import com.demoproject.cricketapp.beans.response.PlayerInfoResponse;
 import com.demoproject.cricketapp.beans.response.TeamInfoResponse;
+import org.springframework.stereotype.Component;
 
 import java.util.List;
 
-public interface CricketServicesFacade {
+@Component
+public interface PlayerTeamFacade {
     public Player addPlayer(PlayerRequest playerRequest);
     public Player getPlayerById(String playerId);
     public List<PlayerInfoResponse> getAllPlayersInfo();
@@ -22,8 +22,9 @@ public interface CricketServicesFacade {
     public Team getTeamById(String teamId);
     public List<TeamInfoResponse> getAllTeamsInfo();
 
-    // MatchService methods
-    public Match getMatchById(String matchId);
 
-    public Scoreboard getMatchScoreboard(String matchId);
+    //Player Team methods
+    public Team addPlayerToTeam(String teamId, String playerId);
+    public Team dropPlayerFromTeam(String teamId, String playerId);
+
 }
