+++ src/main/java/com/demoproject/cricketapp/service/PlayerService.java
@@ -3,8 +3,6 @@ package com.demoproject.cricketapp.service;
 import com.demoproject.cricketapp.beans.Player;
 import com.demoproject.cricketapp.beans.request.PlayerRequest;
 import com.demoproject.cricketapp.beans.response.PlayerInfoResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
