+++ src/main/java/com/demoproject/cricketapp/utils/ScoreboardUtils.java
@@ -1,0 +1,13 @@
+package com.demoproject.cricketapp.utils;
+
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+import lombok.experimental.UtilityClass;
+import org.springframework.stereotype.Component;
+
+
+
+@UtilityClass
+public class ScoreboardUtils {
+
+}
